# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hawk-Eye Control Plane",
    "description": "Insider & privileged-user fraud detection — FastAPI control plane (ALERT-ONLY).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/alerts": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "List Alerts",
        "operationId": "list_alerts_api_v1_alerts_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "risk_gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Gte"
            }
          },
          {
            "name": "assignee",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assignee"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{alert_id}": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "Get Alert",
        "operationId": "get_alert_api_v1_alerts__alert_id__get",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alert"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/compliance/transfers": {
      "post": {
        "tags": [
          "compliance"
        ],
        "summary": "Record Transfer",
        "operationId": "record_transfer_api_v1_compliance_transfers_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Record Transfer Api V1 Compliance Transfers Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/compliance/data-principal": {
      "post": {
        "tags": [
          "compliance"
        ],
        "summary": "Data Principal Request",
        "operationId": "data_principal_request_api_v1_compliance_data_principal_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataPrincipalRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Data Principal Request Api V1 Compliance Data Principal Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Alert": {
        "properties": {
          "alert_id": {
            "type": "string",
            "title": "Alert Id",
            "examples": [
              "alr_3d7e22"
            ]
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id",
            "description": "= employee_id",
            "examples": [
              "EMP-7f3a"
            ]
          },
          "risk_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "examples": [
              87
            ]
          },
          "severity": {
            "$ref": "#/components/schemas/Severity",
            "examples": [
              "high"
            ]
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "examples": [
              0.82
            ]
          },
          "status": {
            "$ref": "#/components/schemas/AlertStatus",
            "default": "open"
          },
          "created_ts": {
            "type": "string",
            "title": "Created Ts",
            "examples": [
              "2026-06-30T02:41:55Z"
            ]
          },
          "contributing_layers": {
            "items": {
              "$ref": "#/components/schemas/ContributingLayer"
            },
            "type": "array",
            "title": "Contributing Layers"
          },
          "reason_codes": {
            "items": {
              "$ref": "#/components/schemas/ReasonCode"
            },
            "type": "array",
            "title": "Reason Codes"
          },
          "exposure_inr": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Exposure Inr",
            "description": "Integer INR (CONTEXT.md §6)",
            "default": 0,
            "examples": [
              4800000
            ]
          },
          "sla_due_ts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Due Ts",
            "examples": [
              "2026-07-30T02:41:55Z"
            ]
          },
          "pii_tokenized": {
            "type": "boolean",
            "title": "Pii Tokenized",
            "description": "Always True on egress; raw PII never in an alert",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "alert_id",
          "entity_id",
          "risk_score",
          "severity",
          "confidence",
          "created_ts"
        ],
        "title": "Alert",
        "description": "L6 fusion output (Part 24.5b)."
      },
      "AlertPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Alert"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "AlertPage",
        "description": "Paginated, ranked alert queue (GET /alerts)."
      },
      "AlertStatus": {
        "type": "string",
        "enum": [
          "open",
          "assigned",
          "in_review",
          "block_requested",
          "confirmed_fraud",
          "false_positive",
          "inconclusive",
          "closed"
        ],
        "title": "AlertStatus"
      },
...
      "ContributingLayer": {
        "type": "string",
        "enum": [
          "L1_rules",
          "L2_unsupervised",
          "L3_gbdt",
          "L4_sequence",
          "L5_graph"
        ],
        "title": "ContributingLayer"
      },
...
      "DataPrincipalRequestBody": {
        "properties": {
          "request_type": {
            "type": "string",
            "title": "Request Type",
            "description": "access | correction | erasure | grievance"
          },
          "principal_id": {
            "type": "string",
            "title": "Principal Id"
          },
          "under_fraud_hold": {
            "type": "boolean",
            "title": "Under Fraud Hold",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "request_type",
          "principal_id"
        ],
        "title": "DataPrincipalRequestBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReasonCode": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/ReasonSource"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code",
            "description": "Rule code, e.g. NEW_BENEFICIARY_THEN_HIGHVALUE"
          },
          "feature": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature",
            "description": "SHAP feature name"
          },
          "contribution": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contribution",
            "description": "SHAP contribution (signed)"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Human-readable evidence (tokenized)"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "source"
        ],
        "title": "ReasonCode",
        "description": "One assembled reason — rule provenance, SHAP contribution, or graph/sequence evidence."
      },
      "ReasonSource": {
        "type": "string",
        "enum": [
          "rule",
          "shap",
          "graph",
          "sequence"
        ],
        "title": "ReasonSource"
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "Severity"
      },
...
      "TransferRequest": {
        "properties": {
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "destination_jurisdiction": {
            "type": "string",
            "title": "Destination Jurisdiction"
          },
          "tee_attested": {
            "type": "boolean",
            "title": "Tee Attested",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "purpose",
          "destination_jurisdiction"
        ],
        "title": "TransferRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}